# hospital-backend/swagger.yml
openapi: 3.0.0
info:
  title: Hospital Backend API
  description: API for a single hospital node in a federated learning network for breast cancer detection.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /patients/upload:
    post:
      tags: [Patients]
      summary: Upload patient data
      description: Upload a histopathology image and metadata for a new patient.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                patient_metadata:
                  type: string
                  format: json
                  description: JSON string containing patient metadata conforming to PatientMetadata schema.
                  example: '{"patientId": "P123456", "age": 45, "diagnosis": "malignant"}'
                histopathology:
                  type: string
                  format: binary
                  description: Histopathology image file (PNG, JPG, TIF)
      responses:
        '201':
          description: Patient data uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  patient_id:
                    type: string
                    example: "7dfd5f6f-3ef8-4325-bd85-2bb53ffb1463"
                  metadata:
                    $ref: '#/components/schemas/PatientMetadata'
                  file_urls:
                    type: object
                    properties:
                      histopathology:
                        $ref: '#/components/schemas/BlobStorageFile'
                  features_extracted:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: "Patient data uploaded and processed successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /patients/list:
    get:
      tags: [Patients]
      summary: List patients
      description: Get a paginated list of patients for this hospital.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful response
        '500':
          $ref: '#/components/responses/InternalError'
  /patients/{id}:
    get:
      tags: [Patients]
      summary: Get patient details
      description: Retrieve full details for a specific patient by their ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    PatientMetadata:
      type: object
      properties:
        patientId:
          type: string
          description: Unique identifier for the patient
          example: "P123456"
        age:
          type: number
          description: Patient's age in years
          example: 45
        diagnosis:
          type: string
          description: Medical diagnosis result
          enum: ["malignant", "benign"]
          example: "malignant"
      required:
        - patientId
        - age
        - diagnosis
    BlobStorageFile:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: Public URL to access the file on Vercel Blob Storage
          example: "https://storage.example.com/patients/7dfd5f6f/histopathology-1762836125293.encrypted"
        downloadUrl:
          type: string
          format: uri
          description: Download URL with download parameter
          example: "https://storage.example.com/patients/7dfd5f6f/histopathology-1762836125293.encrypted?download=1"
        encrypted:
          type: boolean
          description: Whether the file is encrypted
          example: true
        size:
          type: integer
          description: File size in bytes
          example: 174469
        checksum:
          type: string
          description: SHA-256 checksum of the file
          example: "163d481189f2883161ce31acffbe371c56c0853eac22270ae994d3fbc4ff5074"
  responses:
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    BadRequest:
      description: The request was malformed or invalid.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    InternalError:
      description: An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string